backend/pyproject.toml
[project]
name = "hydroguide-backend"
version = "0.1.0"
description = "HydroGuide - Backend API for hydropower station dimensioning"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "openpyxl>=3.1.0",
    "orjson>=3.10.0",
    "python-multipart>=0.0.18",
    "pyjwt>=2.10.0",
    "sse-starlette>=2.2.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "rich>=13.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pyright]
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"
executionEnvironments = [{ root = "." }]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
